在半精度浮点数（FP16）中，大于 2048 的整数无法保证精确表示。在 2048 到 4096 之间，FP16 的最小间隔是 2。需要做一点调整，确保 fp16 下的 spk encoder 可以正常工作。

+++ Qwen3-TTS-main/qwen_tts/core/models/modeling_qwen3_tts.py
@@ -199,7 +199,7 @@ class AttentiveStatisticsPooling(nn.Module):
 
         if max_len is None:
             max_len = length.max().long().item()  # using arange to generate mask
-        mask = torch.arange(max_len, device=length.device, dtype=length.dtype).expand(
+        mask = torch.arange(max_len, device=length.device, dtype=torch.long).to(length.dtype).expand(
             len(length), max_len
         ) < length.unsqueeze(1)
